+++ Console/Script/Console.cs
@@ -40,6 +40,9 @@ namespace ConsoleUtility
         private bool bVisible = false;
         private int history = -1;
 
+        public static event ConsoleDelegate onConsoleToggle;
+        public delegate void ConsoleDelegate(bool visible);
+
         void OnEnable()
         {
             s_ConsoleData = new ConsoleData();
@@ -306,6 +309,8 @@ namespace ConsoleUtility
                 InputField.ActivateInputField();
                 UpdateLog();
             }
+
+            onConsoleToggle?.Invoke(visible);
         }
 
         public static void ExecuteCommand(string command)