+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2021-XX-XX] - 1.3.0
 
+* Added a `onConsoleToggle` event to hook into when the console visibility changes.
 * Added API for debug views.
 
 ## [2021-01-13] - 1.2.0